moved createAccApplicants + added pwd valid. error
